+++ packages/faucet/src/actions/generate.ts
@@ -1,6 +1,7 @@
 import { Bip39, Random } from "@cosmjs/crypto";
 
 import * as constants from "../constants";
+import { makePathBuilder } from "../pathbuilder";
 import { createWallets } from "../profile";
 
 export async function generate(args: readonly string[]): Promise<void> {
@@ -13,6 +14,9 @@ export async function generate(args: readonly string[]): Promise<void> {
   const mnemonic = Bip39.encode(Random.getBytes(16)).toString();
   console.info(`FAUCET_MNEMONIC="${mnemonic}"`);
 
+  const pathBuilder = makePathBuilder(constants.pathPattern);
+  console.info(`FAUCET_PATH_PATTERN="${constants.pathPattern}"`);
+
   // Log the addresses
-  await createWallets(mnemonic, constants.addressPrefix, constants.concurrency, true);
+  await createWallets(mnemonic, pathBuilder, constants.addressPrefix, constants.concurrency, true, true);
 }