+++ packages/faucet/src/profile.ts
@@ -1,16 +1,18 @@
 import { pathToString } from "@cosmjs/crypto";
-import { makeCosmoshubPath, Secp256k1HdWallet, SigningCosmosClient } from "@cosmjs/launchpad";
+import { Secp256k1HdWallet, SigningCosmosClient } from "@cosmjs/launchpad";
 import { DirectSecp256k1HdWallet, isOfflineDirectSigner, OfflineSigner } from "@cosmjs/proto-signing";
 import { SigningStargateClient } from "@cosmjs/stargate";
 
 import * as constants from "./constants";
+import { PathBuilder } from "./pathbuilder";
 
 export async function createWallets(
   mnemonic: string,
+  pathBuilder: PathBuilder,
   addressPrefix: string,
   numberOfDistributors: number,
-  stargate = true,
-  logging = false,
+  stargate: boolean,
+  logging: boolean,
 ): Promise<ReadonlyArray<readonly [string, OfflineSigner]>> {
   const createWallet = stargate ? DirectSecp256k1HdWallet.fromMnemonic : Secp256k1HdWallet.fromMnemonic;
   const wallets = new Array<readonly [string, OfflineSigner]>();
@@ -18,7 +20,7 @@ export async function createWallets(
   // first account is the token holder
   const numberOfIdentities = 1 + numberOfDistributors;
   for (let i = 0; i < numberOfIdentities; i++) {
-    const path = makeCosmoshubPath(i);
+    const path = pathBuilder(i);
     const wallet = await createWallet(mnemonic, { hdPaths: [path], prefix: addressPrefix });
     const [{ address }] = await wallet.getAccounts();
     if (logging) {